Consolidated all uses of dicom parsing to StorageService where the dicom value chunk size can be specified. This chunk size is now configurable
