+++ src/main/resources/application.conf
@@ -81,6 +81,10 @@ slicebox {
     config = ${slicebox.database.h2}
   }
 
+  stream {
+    chunk-size = 5m
+  }
+
   anonymization {
     purge-empty-keys = false
   }
