+++ src/main/scala/se/nimsa/sbx/box/BoxPollActor.scala
@@ -137,7 +137,7 @@ class BoxPollActor(box: Box,
       val statusCode = response.status.intValue
       if (statusCode >= 200 && statusCode < 300) {
         val source = Source(SourceType.BOX, box.name, box.id)
-        storeDicomData(response.entity.dataBytes.via(Compression.inflate()), source, storage, Contexts.extendedContexts).flatMap { metaData =>
+        storeDicomData(response.entity.dataBytes.via(Compression.inflate()), source, storage, Contexts.extendedContexts, reverseAnonymization = true).flatMap { metaData =>
           val overwrite = !metaData.imageAdded
           boxService.ask(UpdateIncoming(box, transactionImage.transaction.id, transactionImage.image.sequenceNumber, transactionImage.transaction.totalImageCount, metaData.image.id, overwrite)).flatMap {
             case IncomingUpdated(transaction) =>
