+++ src/main/scala/se/nimsa/sbx/dicom/streams/DicomStreamOps.scala
@@ -16,16 +16,21 @@
 
 package se.nimsa.sbx.dicom.streams
 
+import java.awt.RenderingHints
+import java.awt.image.BufferedImage
+import java.io.ByteArrayOutputStream
+import javax.imageio.ImageIO
+
 import akka.actor.Cancellable
-import akka.stream.scaladsl.{Broadcast, Flow, GraphDSL, Merge, Sink, Source => StreamSource}
-import akka.stream.{FlowShape, Materializer, SinkShape}
+import akka.stream.scaladsl.{Broadcast, Flow, GraphDSL, Sink, StreamConverters, Source => StreamSource}
+import akka.stream.{Materializer, SinkShape}
 import akka.util.ByteString
 import akka.{Done, NotUsed}
 import org.dcm4che3.data._
+import org.dcm4che3.imageio.plugins.dcm.DicomImageReadParam
 import org.dcm4che3.io.DicomStreamException
 import se.nimsa.dcm4che.streams.DicomFlows._
 import se.nimsa.dcm4che.streams.DicomModifyFlow._
-import se.nimsa.dcm4che.streams.DicomParseFlow._
 import se.nimsa.dcm4che.streams.DicomParts._
 import se.nimsa.dcm4che.streams.TagPath.TagPathSequence
 import se.nimsa.dcm4che.streams._
@@ -35,7 +40,9 @@ import se.nimsa.sbx.dicom.Contexts.Context
 import se.nimsa.sbx.dicom.DicomHierarchy.Image
 import se.nimsa.sbx.dicom.DicomPropertyValue.{PatientID, PatientName}
 import se.nimsa.sbx.dicom.{Contexts, DicomUtil, ImageAttribute}
+import se.nimsa.sbx.lang.NotFoundException
 import se.nimsa.sbx.metadata.MetaDataProtocol._
+import se.nimsa.sbx.storage.StorageProtocol.ImageInformation
 import se.nimsa.sbx.storage.StorageService
 import se.nimsa.sbx.util.SbxExtensions._
 
@@ -48,36 +55,27 @@ import scala.reflect.ClassTag
   */
 trait DicomStreamOps {
 
-  import DicomStreamOps._
+  import AnonymizationFlow._
+  import HarmonizeAnonymizationFlow._
+  import ReverseAnonymizationFlow._
+  import DicomStreamUtil._
 
   def callAnonymizationService[R: ClassTag](message: Any): Future[R]
   def callMetaDataService[R: ClassTag](message: Any): Future[R]
   def scheduleTask(delay: FiniteDuration)(task: => Unit): Cancellable
-
-  protected def getOrCreateAnonKey(tagValues: Seq[TagValue]): DicomInfoPart => Future[AnonymizationKey] =
-    (info: DicomInfoPart) => callAnonymizationService[AnonymizationKey](GetOrCreateAnonymizationKey(
-      info.patientName, info.patientID, info.patientSex, info.patientBirthDate, info.patientAge, info.studyInstanceUID,
-      info.studyDescription, info.studyID, info.accessionNumber, info.seriesInstanceUID, info.seriesDescription,
-      info.protocolName, info.frameOfReferenceUID, tagValues))
 
   /**
     * Creates a streaming source of anonymized and harmonized DICOM data
     *
     * @param imageId   ID of image to load
     * @param tagValues forced values of attributes as pairs of tag number and string value encoded in UTF-8 (ASCII) format
-    * @param storage   the storage backend (file, runtime, S3 etc)
     * @return a `Source` of anonymized DICOM byte chunks
     */
-  def anonymizedDicomData(imageId: Long, tagValues: Seq[TagValue], storage: StorageService)
+  protected def anonymizedDicomData(imageId: Long, tagValues: Seq[TagValue], storage: StorageService)
                          (implicit materializer: Materializer, ec: ExecutionContext): StreamSource[ByteString, NotUsed] = {
-    val source = storage.fileSource(imageId)
-      .via(parseFlow)
+    val source = storage.dataSource(imageId, None)
     anonymizedDicomDataSource(source, getOrCreateAnonKey(tagValues), tagValues)
   }
-
-  protected def reverseAnonymizationKeysForPatient(implicit ec: ExecutionContext): (PatientName, PatientID) => Future[Seq[AnonymizationKey]] =
-    (patientName: PatientName, patientID: PatientID) => callAnonymizationService[AnonymizationKeys](GetReverseAnonymizationKeysForPatient(patientName.value, patientID.value))
-      .map(_.anonymizationKeys)
 
   /**
     * Store DICOM data from a source of byte chunks and update meta data.
@@ -89,10 +87,10 @@ trait DicomStreamOps {
     * @param reverseAnonymization switch to determined whether reverse anonymization should be carried out or not
     * @return the meta data info stored in the database
     */
-  def storeDicomData(bytesSource: StreamSource[ByteString, _], source: Source, storage: StorageService, contexts: Seq[Context], reverseAnonymization: Boolean = true)
+  protected def storeDicomData(bytesSource: StreamSource[ByteString, _], source: Source, storage: StorageService, contexts: Seq[Context], reverseAnonymization: Boolean)
                     (implicit materializer: Materializer, ec: ExecutionContext): Future[MetaDataAdded] = {
     val tempPath = createTempPath()
-    val sink = dicomDataSink(storage.fileSink(tempPath), reverseAnonymizationKeysForPatient, contexts, reverseAnonymization)
+    val sink = dicomDataSink(storage.fileSink(tempPath), storage.parseFlow(None), reverseAnonymizationKeysForPatient, contexts, reverseAnonymization)
     bytesSource.runWith(sink).flatMap {
       case (_, maybeDataset) => storeDicomData(maybeDataset, source, tempPath, storage)
     }.recover {
@@ -101,15 +99,6 @@ trait DicomStreamOps {
           storage.deleteByName(Seq(tempPath)) // delete temp file once file system has released handle
         }
         throw t
-    }
-  }
-
-  private def storeDicomData(maybeDataset: Option[Attributes], source: Source, tempPath: String, storage: StorageService)
-                            (implicit ec: ExecutionContext): Future[MetaDataAdded] = {
-    val attributes: Attributes = maybeDataset.getOrElse(throw new DicomStreamException("DICOM data has no dataset"))
-    callMetaDataService[MetaDataAdded](AddMetaData(attributes, source)).map { metaDataAdded =>
-      storage.move(tempPath, storage.imageName(metaDataAdded.image.id))
-      metaDataAdded
     }
   }
 
@@ -122,14 +111,14 @@ trait DicomStreamOps {
     * @param storage   the storage backend (file, runtime, S3 etc)
     * @return the anonymized metadata stored in the system
     */
-  def anonymizeData(imageId: Long, tagValues: Seq[TagValue], storage: StorageService)
+  protected def anonymizeData(imageId: Long, tagValues: Seq[TagValue], storage: StorageService)
                    (implicit materializer: Materializer, ec: ExecutionContext): Future[Option[MetaDataAdded]] =
     callMetaDataService[Option[Image]](GetImage(imageId)).flatMap { imageMaybe =>
       imageMaybe.map { image =>
         callMetaDataService[Option[SeriesSource]](GetSourceForSeries(image.seriesId)).map { seriesSourceMaybe =>
           seriesSourceMaybe.map { seriesSource =>
-            val forcedSource = storage.fileSource(imageId)
-              .via(parseFlow)
+            val forcedSource = storage
+              .dataSource(imageId, None)
               .via(modifyFlow(TagModification.contains(TagPath.fromTag(Tag.PatientIdentityRemoved), _ => ByteString("NO"), insert = false)))
               .via(blacklistFilter(Set(TagPath.fromTag(Tag.DeidentificationMethod))))
             val anonymizedSource = anonymizedDicomDataSource(forcedSource, getOrCreateAnonKey(tagValues), tagValues)
@@ -145,7 +134,7 @@ trait DicomStreamOps {
     }.unwrap
 
 
-  def modifyData(imageId: Long, tagModifications: Seq[TagModification], storage: StorageService)
+  protected def modifyData(imageId: Long, tagModifications: Seq[TagModification], storage: StorageService)
                 (implicit materializer: Materializer, ec: ExecutionContext): Future[(MetaDataDeleted, MetaDataAdded)] = {
 
     val futureSourceAndTags =
@@ -160,11 +149,11 @@ trait DicomStreamOps {
       }.unwrap.map(_.getOrElse((None, Seq.empty)))
 
     val tempPath = createTempPath()
-    val sink = dicomDataSink(storage.fileSink(tempPath), reverseAnonymizationKeysForPatient, Contexts.extendedContexts, reverseAnonymization = false)
+    val sink = dicomDataSink(storage.fileSink(tempPath), storage.parseFlow(None), reverseAnonymizationKeysForPatient, Contexts.extendedContexts, reverseAnonymization = false)
 
     val futureModifiedTempFile =
-      storage.fileSource(imageId)
-        .via(parseFlow)
+      storage
+        .dataSource(imageId, None)
         .via(groupLengthDiscardFilter)
         .via(toUndefinedLengthSequences)
         .via(modifyFlow(tagModifications: _*))
@@ -189,192 +178,9 @@ trait DicomStreamOps {
 
   }
 
-}
-
-object DicomStreamOps {
-
-  import AnonymizationFlow._
-  import HarmonizeAnonymizationFlow._
-  import ReverseAnonymizationFlow._
-
-  val encodingTags = Set(Tag.TransferSyntaxUID, Tag.SpecificCharacterSet)
-
-  val tagsToStoreInDB: Set[Int] = {
-    val patientTags = Seq(Tag.PatientName, Tag.PatientID, Tag.PatientSex, Tag.PatientBirthDate)
-    val studyTags = Seq(Tag.StudyInstanceUID, Tag.StudyDescription, Tag.StudyID, Tag.StudyDate, Tag.AccessionNumber, Tag.PatientAge)
-    val seriesTags = Seq(Tag.SeriesInstanceUID, Tag.SeriesDescription, Tag.SeriesDate, Tag.Modality, Tag.ProtocolName, Tag.BodyPartExamined, Tag.Manufacturer, Tag.StationName, Tag.FrameOfReferenceUID)
-    val imageTags = Seq(Tag.SOPInstanceUID, Tag.ImageType, Tag.InstanceNumber)
-
-    encodingTags ++ patientTags ++ studyTags ++ seriesTags ++ imageTags
-  }
-
-  val basicInfoTags: Set[Int] = encodingTags ++ Set(Tag.PatientName, Tag.PatientID, Tag.PatientIdentityRemoved,
-    Tag.StudyInstanceUID, Tag.SeriesInstanceUID)
-
-  val extendedInfoTags: Set[Int] = basicInfoTags ++ Set(Tag.PatientSex, Tag.PatientBirthDate, Tag.PatientAge, Tag.StudyDescription, Tag.StudyID,
-    Tag.AccessionNumber, Tag.SeriesDescription, Tag.ProtocolName, Tag.FrameOfReferenceUID)
-
-  case class DicomInfoPart(transferSyntaxUid: Option[String],
-                           specificCharacterSet: Option[SpecificCharacterSet],
-                           patientID: Option[String],
-                           patientName: Option[String],
-                           patientSex: Option[String],
-                           patientBirthDate: Option[String],
-                           patientAge: Option[String],
-                           identityRemoved: Option[String],
-                           studyInstanceUID: Option[String],
-                           studyDescription: Option[String],
-                           studyID: Option[String],
-                           accessionNumber: Option[String],
-                           seriesInstanceUID: Option[String],
-                           seriesDescription: Option[String],
-                           protocolName: Option[String],
-                           frameOfReferenceUID: Option[String]) extends DicomPart {
-    def bytes: ByteString = ByteString.empty
-    def bigEndian: Boolean = false
-    def isAnonymized: Boolean = identityRemoved.exists(_.toUpperCase == "YES")
-  }
-
-  case class PartialAnonymizationKeyPart(keyMaybe: Option[AnonymizationKey], hasPatientInfo: Boolean, hasStudyInfo: Boolean, hasSeriesInfo: Boolean) extends DicomPart {
-    def bytes: ByteString = ByteString.empty
-    def bigEndian: Boolean = false
-  }
-
-  def maybeDeflateFlow: Flow[DicomPart, DicomPart, NotUsed] = conditionalFlow(
-    {
-      case p: DicomInfoPart => p.transferSyntaxUid.isDefined && DicomParsing.isDeflated(p.transferSyntaxUid.get)
-    }, deflateDatasetFlow, Flow.fromFunction(identity), routeADefault = false)
-
-  def createTempPath() = s"tmp-${java.util.UUID.randomUUID().toString}"
-
-  def attributesToInfoPart(dicomPart: DicomPart)
-                          (implicit ec: ExecutionContext, materializer: Materializer): Future[DicomPart] = {
-    dicomPart match {
-      case da: DicomAttributes =>
-        StreamSource.fromIterator(() => da.attributes.iterator).runWith(DicomAttributesSink.attributesSink).map {
-          case (fmiMaybe, dsMaybe) =>
-            DicomInfoPart(
-              fmiMaybe.flatMap(fmi => Option(fmi.getString(Tag.TransferSyntaxUID))),
-              dsMaybe.flatMap(ds => Option(ds.getSpecificCharacterSet)),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.PatientID))),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.PatientName))),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.PatientSex))),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.PatientBirthDate))),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.PatientAge))),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.PatientIdentityRemoved))),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.StudyInstanceUID))),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.StudyDescription))),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.StudyID))),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.AccessionNumber))),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.SeriesInstanceUID))),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.SeriesDescription))),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.ProtocolName))),
-              dsMaybe.flatMap(ds => Option(ds.getString(Tag.FrameOfReferenceUID))))
-        }
-      case part: DicomPart => Future.successful(part)
-    }
-  }
-
-  def reverseAnonymizationKeyPartForPatient(query: (PatientName, PatientID) => Future[Seq[AnonymizationKey]])
-                                        (implicit ec: ExecutionContext): DicomPart => Future[List[DicomPart]] = {
-    case info: DicomInfoPart =>
-      val maybeFutureParts = for {
-        patientName <- info.patientName if info.isAnonymized
-        patientID <- info.patientID
-      } yield {
-        query(PatientName(patientName), PatientID(patientID)).map { patientKeys =>
-          val studyKeys = info.studyInstanceUID.map(studyUID => patientKeys.filter(_.anonStudyInstanceUID == studyUID)).getOrElse(Seq.empty)
-          val seriesKeys = info.seriesInstanceUID.map(seriesUID => studyKeys.filter(_.anonSeriesInstanceUID == seriesUID)).getOrElse(Seq.empty)
-          val maybeKey = seriesKeys.headOption.orElse(studyKeys.headOption).orElse(patientKeys.headOption)
-          info :: PartialAnonymizationKeyPart(maybeKey, hasPatientInfo = patientKeys.nonEmpty, hasStudyInfo = studyKeys.nonEmpty, hasSeriesInfo = seriesKeys.nonEmpty) :: Nil
-        }
-      }
-      maybeFutureParts.getOrElse(Future.successful(info :: PartialAnonymizationKeyPart(None, hasPatientInfo = false, hasStudyInfo = false, hasSeriesInfo = false) :: Nil))
-    case part: DicomPart =>
-      Future.successful(part :: Nil)
-  }
-
-  def dicomDataSink(storageSink: Sink[ByteString, Future[Done]], reverseAnonymizationKeysForPatient: (PatientName, PatientID) => Future[Seq[AnonymizationKey]], contexts: Seq[Context], reverseAnonymization: Boolean = true)
-                   (implicit ec: ExecutionContext, materializer: Materializer): Sink[ByteString, Future[(Option[Attributes], Option[Attributes])]] = {
-
-    val attributesSink = DicomAttributesSink.attributesSink
-
-    val validationContexts = Contexts.asNamePairs(contexts).map(ValidationContext.tupled)
-
-    def runBothKeepRight[A, B] = (futureLeft: Future[A], futureRight: Future[B]) => futureLeft.flatMap(_ => futureRight)
-
-    Sink.fromGraph(GraphDSL.create(storageSink, attributesSink)(runBothKeepRight) { implicit builder =>
-      (storageSink, attributesSink) =>
-        import GraphDSL.Implicits._
-
-        val baseFlow = validateFlowWithContext(validationContexts, drainIncoming = true)
-          .via(parseFlow)
-          .via(collectAttributesFlow(basicInfoTags))
-          .mapAsync(1)(attributesToInfoPart) // needed for e.g. maybe deflate flow
-
-        val flow = builder.add {
-          if (reverseAnonymization)
-            baseFlow
-              .mapAsync(1)(reverseAnonymizationKeyPartForPatient(reverseAnonymizationKeysForPatient))
-              .mapConcat(identity) // flatten stream of lists
-              .via(maybeReverseAnonFlow)
-          else
-            baseFlow
-        }
-
-        val bcast = builder.add(Broadcast[DicomPart](2))
-
-        flow ~> bcast.in
-        bcast.out(0) ~> maybeDeflateFlow.map(_.bytes) ~> storageSink
-        bcast.out(1) ~> whitelistFilter(tagsToStoreInDB) ~> attributeFlow ~> attributesSink
-
-        SinkShape(flow.in)
-    })
-  }
-
-  def getOrCreateAnonKeyPart(getOrCreateAnonKey: DicomInfoPart => Future[AnonymizationKey])
-                            (implicit ec: ExecutionContext): DicomPart => Future[DicomPart] = {
-    case info: DicomInfoPart if info.isAnonymized =>
-      Future.successful(PartialAnonymizationKeyPart(None, hasPatientInfo = false, hasStudyInfo = false, hasSeriesInfo = false))
-    case info: DicomInfoPart =>
-      getOrCreateAnonKey(info).map(key => PartialAnonymizationKeyPart(Some(key), hasPatientInfo = true, hasStudyInfo = true, hasSeriesInfo = true))
-    case part: DicomPart => Future.successful(part)
-  }
-
-  def toTagModifications(tagValues: Seq[TagValue]): Seq[TagModification] =
-    tagValues.map(tv => TagModification.endsWith(TagPath.fromTag(tv.tag), _ => padToEvenLength(ByteString(tv.value), tv.tag), insert = true))
-
-  def anonymizedDicomDataSource(storageSource: StreamSource[DicomPart, NotUsed],
-                                getOrCreateAnonKey: DicomInfoPart => Future[AnonymizationKey],
-                                tagValues: Seq[TagValue])
-                               (implicit ec: ExecutionContext, materializer: Materializer): StreamSource[ByteString, NotUsed] =
-    storageSource // DicomPart...
-      .via(collectAttributesFlow(extendedInfoTags)) // DicomAttributes :: DicomPart...
-      .mapAsync(1)(attributesToInfoPart) // DicomAnonPart :: DicomPart...
-      .mapAsync(1)(getOrCreateAnonKeyPart(getOrCreateAnonKey)) // AnonymizationKeyPart :: DicomPart...
-      .via(maybeAnonFlow) // DicomAnonPart needed here
-      .via(maybeHarmonizeAnonFlow) // DicomAnonPart and AnonymizationKeysPart needed here
-      .via(modifyFlow(toTagModifications(tagValues): _*))
-      .via(fmiGroupLengthFlow) // update meta information group length
-      .via(maybeDeflateFlow)
-      .map(_.bytes)
-
-  def inflatedSource(source: StreamSource[ByteString, _]): StreamSource[ByteString, _] = source
-    .via(parseFlow)
-    .via(modifyFlow(
-      TagModification.contains(TagPath.fromTag(Tag.TransferSyntaxUID), valueBytes => {
-        valueBytes.utf8String.trim match {
-          case UID.DeflatedExplicitVRLittleEndian => padToEvenLength(ByteString(UID.ExplicitVRLittleEndian), VR.UI)
-          case _ => valueBytes
-        }
-      }, insert = false)))
-    .via(fmiGroupLengthFlow)
-    .map(_.bytes)
-
-  def imageAttributesSource[M](source: StreamSource[ByteString, M])
-                              (implicit ec: ExecutionContext, materializer: Materializer): StreamSource[ImageAttribute, M] =
-    source
-      .via(new DicomParseFlow(stopTag = Some(Tag.PixelData)))
+  protected def readImageAttributes(imageId: Long, storage: StorageService)(implicit materializer: Materializer, ec: ExecutionContext): StreamSource[ImageAttribute, NotUsed] =
+    storage
+      .dataSource(imageId, Some(Tag.PixelData))
       .via(bulkDataFilter)
       .via(collectAttributesFlow(encodingTags))
       .mapAsync(1)(attributesToInfoPart)
@@ -441,35 +247,193 @@ object DicomStreamOps {
         }
       }
 
-  def conditionalFlow(goA: PartialFunction[DicomPart, Boolean], flowA: Flow[DicomPart, DicomPart, _], flowB: Flow[DicomPart, DicomPart, _], routeADefault: Boolean = true): Flow[DicomPart, DicomPart, NotUsed] =
-    Flow.fromGraph(GraphDSL.create() { implicit builder =>
-      import GraphDSL.Implicits._
-
-      var routeA = routeADefault // determines which path is taken in the graph
-
-      // if any part indicates that alternate route should be taken, remember this
-      val gateFlow = builder.add {
-        Flow[DicomPart].map { part =>
-          if (goA.isDefinedAt(part)) routeA = goA(part)
-          part
-        }
+  protected def readImageInformation(imageId: Long, storage: StorageService)(implicit materializer: Materializer, ec: ExecutionContext): Future[ImageInformation] =
+    storage
+      .dataSource(imageId, Some(imageInformationTags.max + 1))
+      .via(DicomFlows.whitelistFilter(imageInformationTags))
+      .via(DicomFlows.attributeFlow)
+      .runWith(DicomAttributesSink.attributesSink)
+      .map {
+        case (_, maybeAttributes) =>
+          maybeAttributes.map { attributes =>
+            val instanceNumber = attributes.getInt(Tag.InstanceNumber, 1)
+            val imageIndex = attributes.getInt(Tag.ImageIndex, 1)
+            val frameIndex = if (instanceNumber > imageIndex) instanceNumber else imageIndex
+            ImageInformation(
+              attributes.getInt(Tag.NumberOfFrames, 1),
+              frameIndex,
+              attributes.getInt(Tag.SmallestImagePixelValue, 0),
+              attributes.getInt(Tag.LargestImagePixelValue, 0))
+          }.getOrElse(ImageInformation(0, 0, 0, 0))
       }
 
-      // split the flow
-      val bcast = builder.add(Broadcast[DicomPart](2))
+  protected def readPngImageData(imageId: Long, frameNumber: Int, windowMin: Int, windowMax: Int, imageHeight: Int, storage: StorageService)(implicit materializer: Materializer, ec: ExecutionContext): Future[Array[Byte]] = Future {
+    // dcm4che does not support viewing of deflated data, cf. Github issue #42
+    // As a workaround, do streaming inflate and mapping of transfer syntax
+    val source = inflatedSource(storage.dataSource(imageId, None))
+    val is = source.runWith(StreamConverters.asInputStream())
+    val iis = ImageIO.createImageInputStream(is)
 
-      // define gates for each path, only one path is used
-      val gateA = Flow[DicomPart].filter(_ => routeA)
-      val gateB = Flow[DicomPart].filterNot(_ => routeA)
+    val imageReader = ImageIO.getImageReadersByFormatName("DICOM").next
+    imageReader.setInput(iis)
+    val param = imageReader.getDefaultReadParam.asInstanceOf[DicomImageReadParam]
+    if (windowMin < windowMax) {
+      param.setWindowCenter((windowMax - windowMin) / 2)
+      param.setWindowWidth(windowMax - windowMin)
+    }
 
-      // merge the two paths
-      val merge = builder.add(Merge[DicomPart](2))
+    try {
+      val bi = try {
+        val image = imageReader.read(frameNumber - 1, param)
+        scaleImage(image, imageHeight)
+      } catch {
+        case e: NotFoundException => throw e
+        case e: Exception => throw new IllegalArgumentException(e)
+      }
+      val baos = new ByteArrayOutputStream
+      ImageIO.write(bi, "png", baos)
+      baos.close()
+      baos.toByteArray
+    } finally {
+      iis.close()
+    }
+  }
 
-      // surround each flow by gates, remember that flows may produce items without input
-      gateFlow ~> bcast ~> gateA ~> flowA ~> gateA ~> merge
-      bcast ~> gateB ~> flowB ~> gateB ~> merge
+  private[streams] def reverseAnonymizationKeysForPatient(implicit ec: ExecutionContext): (PatientName, PatientID) => Future[Seq[AnonymizationKey]] =
+    (patientName: PatientName, patientID: PatientID) => callAnonymizationService[AnonymizationKeys](GetReverseAnonymizationKeysForPatient(patientName.value, patientID.value))
+      .map(_.anonymizationKeys)
 
-      FlowShape(gateFlow.in, merge.out)
+  private[streams] def getOrCreateAnonKey(tagValues: Seq[TagValue]): DicomInfoPart => Future[AnonymizationKey] =
+    (info: DicomInfoPart) => callAnonymizationService[AnonymizationKey](GetOrCreateAnonymizationKey(
+      info.patientName, info.patientID, info.patientSex, info.patientBirthDate, info.patientAge, info.studyInstanceUID,
+      info.studyDescription, info.studyID, info.accessionNumber, info.seriesInstanceUID, info.seriesDescription,
+      info.protocolName, info.frameOfReferenceUID, tagValues))
+
+  private[streams] def scaleImage(image: BufferedImage, imageHeight: Int): BufferedImage = {
+    val ratio = imageHeight / image.getHeight.asInstanceOf[Double]
+    if (ratio != 0.0 && ratio != 1.0) {
+      val imageWidth = (image.getWidth * ratio).asInstanceOf[Int]
+      val resized = new BufferedImage(imageWidth, imageHeight, image.getType)
+      val g = resized.createGraphics()
+      g.setRenderingHint(RenderingHints.KEY_INTERPOLATION, RenderingHints.VALUE_INTERPOLATION_BILINEAR)
+      g.drawImage(image, 0, 0, imageWidth, imageHeight, 0, 0, image.getWidth, image.getHeight, null)
+      g.dispose()
+      resized
+    } else {
+      image
+    }
+
+  }
+
+  private[streams] def storeDicomData(maybeDataset: Option[Attributes], source: Source, tempPath: String, storage: StorageService)
+                                     (implicit ec: ExecutionContext): Future[MetaDataAdded] = {
+    val attributes: Attributes = maybeDataset.getOrElse(throw new DicomStreamException("DICOM data has no dataset"))
+    callMetaDataService[MetaDataAdded](AddMetaData(attributes, source)).map { metaDataAdded =>
+      storage.move(tempPath, storage.imageName(metaDataAdded.image.id))
+      metaDataAdded
+    }
+  }
+
+  private[streams] def maybeDeflateFlow: Flow[DicomPart, DicomPart, NotUsed] = conditionalFlow(
+    {
+      case p: DicomInfoPart => p.transferSyntaxUid.isDefined && DicomParsing.isDeflated(p.transferSyntaxUid.get)
+    }, deflateDatasetFlow, Flow.fromFunction(identity), routeADefault = false)
+
+  private[streams] def createTempPath() = s"tmp-${java.util.UUID.randomUUID().toString}"
+
+  private[streams] def reverseAnonymizationKeyPartForPatient(query: (PatientName, PatientID) => Future[Seq[AnonymizationKey]])
+                                           (implicit ec: ExecutionContext): DicomPart => Future[List[DicomPart]] = {
+    case info: DicomInfoPart =>
+      val maybeFutureParts = for {
+        patientName <- info.patientName if info.isAnonymized
+        patientID <- info.patientID
+      } yield {
+        query(PatientName(patientName), PatientID(patientID)).map { patientKeys =>
+          val studyKeys = info.studyInstanceUID.map(studyUID => patientKeys.filter(_.anonStudyInstanceUID == studyUID)).getOrElse(Seq.empty)
+          val seriesKeys = info.seriesInstanceUID.map(seriesUID => studyKeys.filter(_.anonSeriesInstanceUID == seriesUID)).getOrElse(Seq.empty)
+          val maybeKey = seriesKeys.headOption.orElse(studyKeys.headOption).orElse(patientKeys.headOption)
+          info :: PartialAnonymizationKeyPart(maybeKey, hasPatientInfo = patientKeys.nonEmpty, hasStudyInfo = studyKeys.nonEmpty, hasSeriesInfo = seriesKeys.nonEmpty) :: Nil
+        }
+      }
+      maybeFutureParts.getOrElse(Future.successful(info :: PartialAnonymizationKeyPart(None, hasPatientInfo = false, hasStudyInfo = false, hasSeriesInfo = false) :: Nil))
+    case part: DicomPart =>
+      Future.successful(part :: Nil)
+  }
+
+  private[streams] def dicomDataSink(storageSink: Sink[ByteString, Future[Done]], parseFlow: DicomParseFlow, reverseAnonymizationKeysForPatient: (PatientName, PatientID) => Future[Seq[AnonymizationKey]], contexts: Seq[Context], reverseAnonymization: Boolean)
+                   (implicit ec: ExecutionContext, materializer: Materializer): Sink[ByteString, Future[(Option[Attributes], Option[Attributes])]] = {
+
+    val attributesSink = DicomAttributesSink.attributesSink
+
+    val validationContexts = Contexts.asNamePairs(contexts).map(ValidationContext.tupled)
+
+    def runBothKeepRight[A, B] = (futureLeft: Future[A], futureRight: Future[B]) => futureLeft.flatMap(_ => futureRight)
+
+    Sink.fromGraph(GraphDSL.create(storageSink, attributesSink)(runBothKeepRight) { implicit builder =>
+      (storageSink, attributesSink) =>
+        import GraphDSL.Implicits._
+
+        val baseFlow = validateFlowWithContext(validationContexts, drainIncoming = true)
+          .via(parseFlow)
+          .via(collectAttributesFlow(basicInfoTags))
+          .mapAsync(1)(attributesToInfoPart) // needed for e.g. maybe deflate flow
+
+        val flow = builder.add {
+          if (reverseAnonymization)
+            baseFlow
+              .mapAsync(1)(reverseAnonymizationKeyPartForPatient(reverseAnonymizationKeysForPatient))
+              .mapConcat(identity) // flatten stream of lists
+              .via(maybeReverseAnonFlow)
+          else
+            baseFlow
+        }
+
+        val bcast = builder.add(Broadcast[DicomPart](2))
+
+        flow ~> bcast.in
+        bcast.out(0) ~> maybeDeflateFlow.map(_.bytes) ~> storageSink
+        bcast.out(1) ~> whitelistFilter(tagsToStoreInDB) ~> attributeFlow ~> attributesSink
+
+        SinkShape(flow.in)
     })
+  }
+
+  private[streams] def getOrCreateAnonKeyPart(getOrCreateAnonKey: DicomInfoPart => Future[AnonymizationKey])
+                            (implicit ec: ExecutionContext): DicomPart => Future[DicomPart] = {
+    case info: DicomInfoPart if info.isAnonymized =>
+      Future.successful(PartialAnonymizationKeyPart(None, hasPatientInfo = false, hasStudyInfo = false, hasSeriesInfo = false))
+    case info: DicomInfoPart =>
+      getOrCreateAnonKey(info).map(key => PartialAnonymizationKeyPart(Some(key), hasPatientInfo = true, hasStudyInfo = true, hasSeriesInfo = true))
+    case part: DicomPart => Future.successful(part)
+  }
+
+  private[streams] def toTagModifications(tagValues: Seq[TagValue]): Seq[TagModification] =
+    tagValues.map(tv => TagModification.endsWith(TagPath.fromTag(tv.tag), _ => padToEvenLength(ByteString(tv.value), tv.tag), insert = true))
+
+  private[streams] def anonymizedDicomDataSource(storageSource: StreamSource[DicomPart, NotUsed],
+                                getOrCreateAnonKey: DicomInfoPart => Future[AnonymizationKey],
+                                tagValues: Seq[TagValue])
+                               (implicit ec: ExecutionContext, materializer: Materializer): StreamSource[ByteString, NotUsed] =
+    storageSource // DicomPart...
+      .via(collectAttributesFlow(extendedInfoTags)) // DicomAttributes :: DicomPart...
+      .mapAsync(1)(attributesToInfoPart) // DicomAnonPart :: DicomPart...
+      .mapAsync(1)(getOrCreateAnonKeyPart(getOrCreateAnonKey)) // AnonymizationKeyPart :: DicomPart...
+      .via(maybeAnonFlow) // DicomAnonPart needed here
+      .via(maybeHarmonizeAnonFlow) // DicomAnonPart and AnonymizationKeysPart needed here
+      .via(modifyFlow(toTagModifications(tagValues): _*))
+      .via(fmiGroupLengthFlow) // update meta information group length
+      .via(maybeDeflateFlow)
+      .map(_.bytes)
+
+  private[streams] def inflatedSource(source: StreamSource[DicomPart, NotUsed]): StreamSource[ByteString, _] = source
+    .via(modifyFlow(
+      TagModification.contains(TagPath.fromTag(Tag.TransferSyntaxUID), valueBytes => {
+        valueBytes.utf8String.trim match {
+          case UID.DeflatedExplicitVRLittleEndian => padToEvenLength(ByteString(UID.ExplicitVRLittleEndian), VR.UI)
+          case _ => valueBytes
+        }
+      }, insert = false)))
+    .via(fmiGroupLengthFlow)
+    .map(_.bytes)
 
 }
