+++ src/test/scala/se/nimsa/sbx/dicom/streams/DicomStreamOpsTest.scala
@@ -14,12 +14,11 @@ import org.dcm4che3.io.{DicomOutputStream, DicomStreamException}
 import org.scalatest.{AsyncFlatSpecLike, BeforeAndAfterAll, BeforeAndAfterEach, Matchers}
 import se.nimsa.dcm4che.streams.DicomFlows.attributeFlow
 import se.nimsa.dcm4che.streams.DicomModifyFlow.TagModification
-import se.nimsa.dcm4che.streams.DicomParseFlow.parseFlow
 import se.nimsa.dcm4che.streams._
 import se.nimsa.sbx.anonymization.AnonymizationProtocol._
 import se.nimsa.sbx.anonymization.{AnonymizationDAO, AnonymizationServiceActor}
 import se.nimsa.sbx.app.GeneralProtocol.{Source, SourceType}
-import se.nimsa.sbx.dicom.streams.DicomStreamOps.{DicomInfoPart, PartialAnonymizationKeyPart}
+import se.nimsa.sbx.dicom.streams.DicomStreamUtil._
 import se.nimsa.sbx.dicom.{Contexts, DicomData}
 import se.nimsa.sbx.metadata.MetaDataProtocol._
 import se.nimsa.sbx.metadata.{MetaDataDAO, MetaDataServiceActor, PropertiesDAO}
@@ -62,16 +61,25 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     await(anonymizationDao.clear())
   }
 
+  val storage = new RuntimeStorage
+
   class DicomStreamOpsImpl extends DicomStreamOps {
 
     override def callAnonymizationService[R: ClassTag](message: Any): Future[R] = anonymizationService.ask(message).mapTo[R]
     override def callMetaDataService[R: ClassTag](message: Any): Future[R] = metaDataService.ask(message).mapTo[R]
     override def scheduleTask(delay: FiniteDuration)(task: => Unit): Cancellable = system.scheduler.scheduleOnce(delay)(task)
+
+    def storeDicomData(bytesSource: StreamSource[ByteString, _], source: Source): Future[MetaDataAdded] =
+      storeDicomData(bytesSource, source, storage, Contexts.imageDataContexts, reverseAnonymization = true)
+
+    def anonymizedDicomData(imageId: Long, tagValues: Seq[TagValue]): StreamSource[ByteString, NotUsed] =
+      anonymizedDicomData(imageId, tagValues, storage)
+
+    def modifyData(imageId: Long, tagModifications: Seq[TagModification]): Future[(MetaDataDeleted, MetaDataAdded)] =
+      modifyData(imageId, tagModifications, storage)
   }
 
   val dicomStreamOpsImpl = new DicomStreamOpsImpl()
-
-  val storage = new RuntimeStorage
 
   def toSource(dicomData: DicomData): StreamSource[ByteString, NotUsed] = {
     val baos = new ByteArrayOutputStream()
@@ -85,14 +93,14 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val bytes = preamble ++ fmiGroupLength(unsupportedMediaStorageSOPClassUID ++ tsuidExplicitLE) ++ unsupportedMediaStorageSOPClassUID ++ tsuidExplicitLE ++ patientNameJohnDoe
     val source = StreamSource.single(bytes)
     recoverToSucceededIf[DicomStreamException] {
-      source.runWith(DicomStreamOps.dicomDataSink(storage.fileSink("name"), (_, _) => Future.successful(Seq.empty), Contexts.imageDataContexts))
+      source.runWith(dicomStreamOpsImpl.dicomDataSink(storage.fileSink("name"), storage.parseFlow(None), (_, _) => Future.successful(Seq.empty), Contexts.imageDataContexts, reverseAnonymization = true))
     }
   }
 
   it should "pass a supported context" in {
     val bytes = preamble ++ fmiGroupLength(supportedMediaStorageSOPClassUID ++ tsuidExplicitLE) ++ supportedMediaStorageSOPClassUID ++ tsuidExplicitLE ++ patientNameJohnDoe
     val source = StreamSource.single(bytes)
-    source.runWith(DicomStreamOps.dicomDataSink(storage.fileSink("name"), (_, _) => Future.successful(Seq.empty), Contexts.imageDataContexts))
+    source.runWith(dicomStreamOpsImpl.dicomDataSink(storage.fileSink("name"), storage.parseFlow(None), (_, _) => Future.successful(Seq.empty), Contexts.imageDataContexts, reverseAnonymization = true))
       .map(_ => succeed)
   }
 
@@ -100,14 +108,14 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val bytes = preamble ++ fmiGroupLength(unknownMediaStorageSOPClassUID ++ tsuidExplicitLE) ++ unknownMediaStorageSOPClassUID ++ tsuidExplicitLE ++ patientNameJohnDoe
     val source = StreamSource.single(bytes)
     recoverToSucceededIf[DicomStreamException] {
-      source.runWith(DicomStreamOps.dicomDataSink(storage.fileSink("name"), (_, _) => Future.successful(Seq.empty), Contexts.imageDataContexts))
+      source.runWith(dicomStreamOpsImpl.dicomDataSink(storage.fileSink("name"), storage.parseFlow(None), (_, _) => Future.successful(Seq.empty), Contexts.imageDataContexts, reverseAnonymization = true))
     }
   }
 
   it should "accept DICOM data with missing file meta information" in {
     val bytes = supportedSOPClassUID
     val source = StreamSource.single(bytes)
-    source.runWith(DicomStreamOps.dicomDataSink(storage.fileSink("name"), (_, _) => Future.successful(Seq.empty), Contexts.imageDataContexts))
+    source.runWith(dicomStreamOpsImpl.dicomDataSink(storage.fileSink("name"), storage.parseFlow(None), (_, _) => Future.successful(Seq.empty), Contexts.imageDataContexts, reverseAnonymization = true))
       .map(_ => succeed)
   }
 
@@ -116,13 +124,13 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val t2 = TagValue(Tag.PatientID, "Mapped Patient ID")
     val t3 = TagValue(Tag.SeriesDescription, "Mapped Series Description")
     val bytes = preamble ++ fmiGroupLength(supportedMediaStorageSOPClassUID ++ tsuidExplicitLE) ++ supportedMediaStorageSOPClassUID ++ tsuidExplicitLE ++ patientNameJohnDoe
-    val source = StreamSource.single(bytes).via(parseFlow)
-    val anonSource = DicomStreamOps.anonymizedDicomDataSource(
+    val source = StreamSource.single(bytes).via(storage.parseFlow(None))
+    val anonSource = dicomStreamOpsImpl.anonymizedDicomDataSource(
       source,
       _ => Future.successful(TestUtil.createAnonymizationKey(new Attributes())),
       Seq(t1, t2, t3))
 
-    anonSource.via(parseFlow).via(DicomFlows.attributeFlow).runWith(DicomAttributesSink.attributesSink).map {
+    anonSource.via(storage.parseFlow(None)).via(DicomFlows.attributeFlow).runWith(DicomAttributesSink.attributesSink).map {
       case (_, dsMaybe) =>
         val ds = dsMaybe.get
         ds.getString(Tag.PatientName) should be("Mapped Patient Name")
@@ -135,10 +143,10 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val attributes = createAttributes
     val key = TestUtil.createAnonymizationKey(attributes, anonPatientName = "apn", anonPatientID = "apid", anonSeriesInstanceUID = "aseuid")
 
-    val source = toSource(DicomData(attributes, null)).via(parseFlow)
-    val anonSource = DicomStreamOps.anonymizedDicomDataSource(source, _ => Future.successful(key), Seq.empty)
+    val source = toSource(DicomData(attributes, null)).via(storage.parseFlow(None))
+    val anonSource = dicomStreamOpsImpl.anonymizedDicomDataSource(source, _ => Future.successful(key), Seq.empty)
 
-    anonSource.via(parseFlow).via(DicomFlows.attributeFlow).runWith(DicomAttributesSink.attributesSink).map {
+    anonSource.via(storage.parseFlow(None)).via(DicomFlows.attributeFlow).runWith(DicomAttributesSink.attributesSink).map {
       case (_, dsMaybe) =>
         val ds = dsMaybe.get
         ds.getString(Tag.PatientName) should be("apn")
@@ -152,11 +160,11 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
   "The DICOM data source with anonymization" should "anonymize the patient ID and harmonize it according to an anonymization key" in {
     val dicomData = TestUtil.createDicomData()
     val anonKey = TestUtil.createAnonymizationKey(dicomData.attributes)
-    val source = toSource(dicomData).via(parseFlow)
+    val source = toSource(dicomData).via(storage.parseFlow(None))
 
-    val anonSource = DicomStreamOps.anonymizedDicomDataSource(source, _ => Future.successful(anonKey), Seq.empty)
+    val anonSource = dicomStreamOpsImpl.anonymizedDicomDataSource(source, _ => Future.successful(anonKey), Seq.empty)
 
-    anonSource.via(parseFlow).via(DicomFlows.attributeFlow).runWith(DicomAttributesSink.attributesSink).map {
+    anonSource.via(storage.parseFlow(None)).via(DicomFlows.attributeFlow).runWith(DicomAttributesSink.attributesSink).map {
       case (_, dsMaybe) =>
         val ds = dsMaybe.get
         ds.getString(Tag.PatientID) shouldBe anonKey.anonPatientID
@@ -173,7 +181,7 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val storedImageIds = Future.sequence {
       sopInstanceUIDs.map { sopInstanceUID =>
         val modifiedBytesSource = bytesSource
-          .via(parseFlow)
+          .via(storage.parseFlow(None))
           .via(DicomFlows.blacklistFilter(Set(TagPath.fromTag(Tag.PixelData))))
           .via(DicomModifyFlow.modifyFlow(
             TagModification.contains(
@@ -185,7 +193,7 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
               uid => uid.dropRight(3) ++ ByteString(f"$sopInstanceUID%03d"),
               insert = true)))
           .map(_.bytes)
-        dicomStreamOpsImpl.storeDicomData(modifiedBytesSource, source, storage, Contexts.imageDataContexts, reverseAnonymization = false)
+        dicomStreamOpsImpl.storeDicomData(modifiedBytesSource, source)
           .map(_.image.id)
       }
     }
@@ -194,8 +202,8 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val anonymousPatientIds = storedImageIds.flatMap { imageIds =>
       Future.sequence {
         imageIds.map { imageId =>
-          dicomStreamOpsImpl.anonymizedDicomData(imageId, Seq.empty, storage)
-            .via(parseFlow)
+          dicomStreamOpsImpl.anonymizedDicomData(imageId, Seq.empty)
+            .via(storage.parseFlow(None))
             .via(attributeFlow)
             .runWith(DicomAttributesSink.attributesSink)
             .map(_._2)
@@ -213,7 +221,7 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val key = TestUtil.createAnonymizationKey(attributes, anonPatientName = "apn", anonPatientID = "apid", anonStudyInstanceUID = "astuid", anonSeriesInstanceUID = "aseuid")
 
     val metaData = DicomInfoPart(None, None, Some("apid"), Some("apn"), None, None, None, Some("YES"), Some("astuid"), None, None, None, Some("aseuid"), None, None, None)
-    val query = DicomStreamOps.reverseAnonymizationKeyPartForPatient((_, _) => Future.successful(Seq(key)))
+    val query = dicomStreamOpsImpl.reverseAnonymizationKeyPartForPatient((_, _) => Future.successful(Seq(key)))
     query(metaData).map { parts =>
       parts should have length 2
       parts.head shouldBe a[DicomInfoPart]
@@ -231,7 +239,7 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val key = TestUtil.createAnonymizationKey(attributes, anonPatientName = "apn", anonPatientID = "apid", anonStudyInstanceUID = "astuid2", anonSeriesInstanceUID = "aseuid")
 
     val metaData = DicomInfoPart(None, None, Some("apid"), Some("apn"), None, None, None, Some("YES"), Some("astuid"), None, None, None, Some("aseuid"), None, None, None)
-    val query = DicomStreamOps.reverseAnonymizationKeyPartForPatient((_, _) => Future.successful(Seq(key)))
+    val query = dicomStreamOpsImpl.reverseAnonymizationKeyPartForPatient((_, _) => Future.successful(Seq(key)))
     query(metaData).map { parts =>
       parts should have length 2
       parts.head shouldBe a[DicomInfoPart]
@@ -249,7 +257,7 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val key = TestUtil.createAnonymizationKey(attributes, anonPatientName = "apn", anonPatientID = "apid", anonStudyInstanceUID = "astuid2", anonSeriesInstanceUID = "aseuid")
 
     val metaData = DicomInfoPart(None, None, Some("apid"), Some("apn"), None, None, None, Some("NO"), Some("astuid"), None, None, None, Some("aseuid"), None, None, None)
-    val query = DicomStreamOps.reverseAnonymizationKeyPartForPatient((_, _) => Future.successful(Seq(key)))
+    val query = dicomStreamOpsImpl.reverseAnonymizationKeyPartForPatient((_, _) => Future.successful(Seq(key)))
     query(metaData).map { parts =>
       parts should have length 2
       parts.head shouldBe a[DicomInfoPart]
@@ -263,7 +271,7 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val source = Source(SourceType.USER, "Jane", -1)
     val testData = TestUtil.testImageDicomData()
     val bytesSource = StreamSource.single(ByteString.fromArray(TestUtil.toByteArray(testData)))
-    dicomStreamOpsImpl.storeDicomData(bytesSource, source, storage, Contexts.imageDataContexts, reverseAnonymization = false).map { metaDataAdded =>
+    dicomStreamOpsImpl.storeDicomData(bytesSource, source).map { metaDataAdded =>
       storage.storage.get(storage.imageName(metaDataAdded.image.id)) shouldBe defined
       metaDataAdded.patient.patientName.value shouldBe testData.attributes.getString(Tag.PatientName)
       metaDataAdded.patient.patientID.value shouldBe testData.attributes.getString(Tag.PatientID)
@@ -286,7 +294,7 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     ).copy(patientName = realName, patientID = realID)
     dicomStreamOpsImpl.callAnonymizationService[AnonymizationKeyAdded](AddAnonymizationKey(anonKey)).flatMap { _ =>
       val bytesSource = StreamSource.single(ByteString.fromArray(TestUtil.toByteArray(testData)))
-      dicomStreamOpsImpl.storeDicomData(bytesSource, source, storage, Contexts.imageDataContexts).map { metaDataAdded =>
+      dicomStreamOpsImpl.storeDicomData(bytesSource, source).map { metaDataAdded =>
         metaDataAdded.patient.patientName.value shouldBe realName
         metaDataAdded.patient.patientID.value shouldBe realID
       }
@@ -297,7 +305,7 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val source = Source(SourceType.USER, "Jane", -1)
     val testData = TestUtil.testImageDicomData()
     val bytesSource = StreamSource.single(ByteString.fromArray(TestUtil.toByteArray(testData)))
-    dicomStreamOpsImpl.storeDicomData(bytesSource, source, storage, Contexts.imageDataContexts).map { metaDataAdded =>
+    dicomStreamOpsImpl.storeDicomData(bytesSource, source).map { metaDataAdded =>
       metaDataAdded.patient.patientName.value shouldBe testData.attributes.getString(Tag.PatientName)
     }
   }
@@ -307,8 +315,8 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val testData = TestUtil.testImageDicomData()
     val bytesSource = StreamSource.single(ByteString.fromArray(TestUtil.toByteArray(testData)))
     for {
-      metaDataAdded1 <- dicomStreamOpsImpl.storeDicomData(bytesSource, source, storage, Contexts.imageDataContexts)
-      (_, metaDataAdded2) <- dicomStreamOpsImpl.modifyData(metaDataAdded1.image.id, Seq.empty, storage)
+      metaDataAdded1 <- dicomStreamOpsImpl.storeDicomData(bytesSource, source)
+      (_, metaDataAdded2) <- dicomStreamOpsImpl.modifyData(metaDataAdded1.image.id, Seq.empty)
     } yield {
       storage.storage.get(storage.imageName(metaDataAdded1.image.id)) should not be defined
       storage.storage.get(storage.imageName(metaDataAdded2.image.id)) shouldBe defined
@@ -321,8 +329,8 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val testData = TestUtil.testImageDicomData()
     val bytesSource = StreamSource.single(ByteString.fromArray(TestUtil.toByteArray(testData)))
     for {
-      metaDataAdded1 <- dicomStreamOpsImpl.storeDicomData(bytesSource, source, storage, Contexts.imageDataContexts)
-      (_, metaDataAdded2) <- dicomStreamOpsImpl.modifyData(metaDataAdded1.image.id, Seq(TagModification.endsWith(TagPath.fromTag(Tag.PatientName), _ => ByteString(newName), insert = true)), storage)
+      metaDataAdded1 <- dicomStreamOpsImpl.storeDicomData(bytesSource, source)
+      (_, metaDataAdded2) <- dicomStreamOpsImpl.modifyData(metaDataAdded1.image.id, Seq(TagModification.endsWith(TagPath.fromTag(Tag.PatientName), _ => ByteString(newName), insert = true)))
     } yield {
       metaDataAdded2.patient.patientName.value shouldBe newName
     }
@@ -333,9 +341,9 @@ class DicomStreamOpsTest extends TestKit(ActorSystem("AnonymizationFlowSpec")) w
     val testData = TestUtil.testImageDicomData()
     val bytesSource = StreamSource.single(ByteString.fromArray(TestUtil.toByteArray(testData)))
     for {
-      metaDataAdded1 <- dicomStreamOpsImpl.storeDicomData(bytesSource, source, storage, Contexts.imageDataContexts)
+      metaDataAdded1 <- dicomStreamOpsImpl.storeDicomData(bytesSource, source)
       _ <- dicomStreamOpsImpl.callMetaDataService[SeriesTagAddedToSeries](AddSeriesTagToSeries(SeriesTag(42, "tag1"), metaDataAdded1.series.id))
-      (_, metaDataAdded2) <- dicomStreamOpsImpl.modifyData(metaDataAdded1.image.id, Seq.empty, storage)
+      (_, metaDataAdded2) <- dicomStreamOpsImpl.modifyData(metaDataAdded1.image.id, Seq.empty)
       oldSource <- dicomStreamOpsImpl.callMetaDataService[Option[SeriesSource]](GetSourceForSeries(metaDataAdded1.series.id))
       newSource <- dicomStreamOpsImpl.callMetaDataService[Option[SeriesSource]](GetSourceForSeries(metaDataAdded2.series.id))
       oldSeriesTags <- dicomStreamOpsImpl.callMetaDataService[SeriesTags](GetSeriesTagsForSeries(metaDataAdded1.series.id))
